+++ app/src/main/java/adapters/UserFeedAdapter.java
@@ -1,9 +1,7 @@
-package com.example.sc2infoapp;
+package adapters;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.os.Build;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,15 +9,17 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.RequiresApi;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.bitmap.CircleCrop;
+import interfaces.IPublished;
+import models.Post;
+import com.example.sc2infoapp.R;
+import models.Tournament;
+import models.UserTournament;
 import com.parse.ParseException;
 import com.parse.ParseFile;
-
-import org.w3c.dom.Text;
 
 import java.util.List;
