+++ app/src/main/java/com/example/sc2infoapp/MainActivity.java
@@ -36,9 +36,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        Intent i = new Intent(MainActivity.this, PlayerActivity.class);
-        i.putExtra("playerName", "Serral");
-        startActivity(i);
+
 
         bottomNavigationView = findViewById(R.id.bottom_navigation);
         bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
@@ -65,6 +63,6 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        //bottomNavigationView.setSelectedItemId(R.id.action_home);
+        bottomNavigationView.setSelectedItemId(R.id.action_home);
     }
 }
