+++ app/src/main/java/interfaces/IMatch.java
@@ -1,4 +1,4 @@
-package com.example.sc2infoapp;
+package interfaces;
 
 import java.util.Date;
