+++ app/src/main/java/models/TaskRunner.java
@@ -1,4 +1,4 @@
-package com.example.sc2infoapp;
+package models;
 
 import android.os.Handler;
 import android.os.Looper;
